webmaster_bot/requirements.txt
# Telegram Bot Framework
aiogram==3.22.0

# HTTP клиент
aiohttp>=3.9.0
requests>=2.31.0

# База данных
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# Утилиты
python-dotenv>=1.0.0
pandas>=2.1.0

# Логирование
python-json-logger>=2.0.0
colorlog>=6.8.0

# Дата и время
python-dateutil>=2.8.0

# Excel экспорты
openpyxl>=3.1.0

# Кэширование
aiocache>=0.12.2
